Add support for mapped tasks, add shared modal component
